FIX: Anaconda and PYPI automatic deployment (#109)
PYPI was not accepting sdist uploads based on incorrect version format. Version format was being taken directly from git describe output, giving version numbers such as 1.0.3-4-ga36bedb, which are not valid under the PEP 440 version number specification (https://www.python.org/dev/peps/pep-0440/). This was preventing upload to PYPI with twine, resulting in HTTP 400 errors.

The new auto-version format still uses git describe but now parses it so that the new version number bumps the tag version by one, and adds a dev specifier to the end which represents the number of commits the build is ahead of the previous tag. For example, if the previous tag was 1.0.3 and the current build is 5 commits ahead of the tag then the new version number will be 1.0.4.dev5

To get this version in the meta.yaml file for the conda build, I now use jinja templating with {% data = load_setup_py_data() %}